backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
firebase-admin==6.1.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.0
python-jose==3.3.0
bcrypt==4.0.1
